Integrate real data to encounters

+++ android/FamiliarStrangers/app/src/main/kotlin/com/villetainio/familiarstrangers/models/Encounter.kt
@@ -1,3 +1,3 @@
 package com.villetainio.familiarstrangers.models
 
-data class Encounter(val name: String = "", val times: Int = 0)
+data class Encounter(val name: String = "", val times: Long = 0)
